+++ src/PowerShell/Commands/GetPartnerCustomerSubscriptionUtilization.cs
@@ -78,7 +78,7 @@ namespace Microsoft.Store.PartnerCenter.PowerShell.Commands
         {
             Scheduler.RunTask(async () =>
             {
-                IPartner partner = await PartnerSession.Instance.ClientFactory.CreatePartnerOperationsAsync();
+                IPartner partner = await PartnerSession.Instance.ClientFactory.CreatePartnerOperationsAsync().ConfigureAwait(false);
 
                 IResourceCollectionEnumerator<ResourceCollection<AzureUtilizationRecord>> enumerator;
                 List<PSAzureUtilizationRecord> records = new List<PSAzureUtilizationRecord>();