+++ src/PowerShell/Commands/SetPartnerAzureSubscription.cs
@@ -42,7 +42,7 @@ namespace Microsoft.Store.PartnerCenter.PowerShell.Commands
             {
                 ISubscriptionClient client = await PartnerSession.Instance.ClientFactory.CreateServiceClientAsync<SubscriptionClient>(
                     new[] { $"{PartnerSession.Instance.Context.Environment.AzureEndpoint}//user_impersonation" },
-                    CustomerId);
+                    CustomerId).ConfigureAwait(false);
 
                 RenamedSubscriptionId value = client.Subscriptions.RenameAsync(
                     SubscriptionId,