+++ src/PowerShell/Commands/SuspendPartnerAzureSubscription.cs
@@ -36,11 +36,11 @@ namespace Microsoft.Store.PartnerCenter.PowerShell.Commands
             {
                 ISubscriptionClient client = await PartnerSession.Instance.ClientFactory.CreateServiceClientAsync<SubscriptionClient>(
                     new[] { $"{PartnerSession.Instance.Context.Environment.AzureEndpoint}//user_impersonation" },
-                    CustomerId);
+                    CustomerId).ConfigureAwait(false);
 
                 CanceledSubscriptionId response = await client.Subscriptions.CancelAsync(
-                    SubscriptionId, 
-                    true, 
+                    SubscriptionId,
+                    true,
                     CancellationToken).ConfigureAwait(false);
 
                 WriteObject(response);